Update flags logic
