+++ cmd/githubAdd.go
@@ -21,7 +21,7 @@ var githubAddCmd = &cobra.Command{
 	Long:  `Prepate github account to be used for Kubefirst installation `,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		fmt.Println("githubAddCmd called")
-		flags, err := processGithubAddCmdFlags(cmd)
+		flags, err := flagset.ProcessGithubAddCmdFlags(cmd)
 		if err != nil {
 			return err
 		}
@@ -70,7 +70,7 @@ var githubAddCmd = &cobra.Command{
 func init() {
 	actionCmd.AddCommand(githubAddCmd)
 	currentCommand := githubAddCmd
-	defineGithubCmdFlags(currentCommand)
+	flagset.DefineGithubCmdFlags(currentCommand)
 	flagset.DefineGlobalFlags(currentCommand)
 
 }
