+++ cmd/githubRemove.go
@@ -21,7 +21,7 @@ var githubRemoveCmd = &cobra.Command{
 	Long:  `TBD`,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		fmt.Println("githubRemove called")
-		flags, err := processGithubAddCmdFlags(cmd)
+		flags, err := flagset.ProcessGithubAddCmdFlags(cmd)
 		if err != nil {
 			return err
 		}
@@ -50,6 +50,6 @@ var githubRemoveCmd = &cobra.Command{
 func init() {
 	actionCmd.AddCommand(githubRemoveCmd)
 	currentCommand := githubRemoveCmd
-	defineGithubCmdFlags(currentCommand)
+	flagset.DefineGithubCmdFlags(currentCommand)
 	flagset.DefineGlobalFlags(currentCommand)
 }
