+++ cmd/githubRemoveValidate.go
@@ -7,6 +7,7 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/kubefirst/kubefirst/internal/flagset"
 	"github.com/kubefirst/kubefirst/internal/githubWrapper"
 	"github.com/spf13/cobra"
 )
@@ -49,6 +50,6 @@ var githubRemoveValidateCmd = &cobra.Command{
 
 func init() {
 	actionCmd.AddCommand(githubRemoveValidateCmd)
-	defineGithubCmdFlags(githubRemoveValidateCmd)
+	flagset.DefineGithubCmdFlags(githubRemoveValidateCmd)
 	githubRemoveValidateCmd.MarkFlagRequired("github.owner")
 }
