+++ internal/flagset/installer.go
@@ -1,4 +1,4 @@
-package cmd
+package flagset
 
 import (
 	"log"
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/viper"
 )
 
-func defineInstallerGenericFlags(currentCommand *cobra.Command) {
+func DefineInstallerGenericFlags(currentCommand *cobra.Command) {
 	//Gewneric Installer flags:
 	currentCommand.Flags().String("cluster-name", "kubefirst", "the cluster name, used to identify resources on cloud provider")
 	currentCommand.Flags().String("admin-email", "", "the email address for the administrator as well as for lets-encrypt certificate emails")
@@ -26,7 +26,7 @@ type InstallerGenericFlags struct {
 	RepoGitops   string //To support forks
 }
 
-func processInstallerGenericFlags(cmd *cobra.Command) (InstallerGenericFlags, error) {
+func ProcessInstallerGenericFlags(cmd *cobra.Command) (InstallerGenericFlags, error) {
 	flags := InstallerGenericFlags{}
 	defer viper.WriteConfig()
 
@@ -71,66 +71,4 @@ func processInstallerGenericFlags(cmd *cobra.Command) (InstallerGenericFlags, er
 	flags.RepoGitops = branchGitOps
 
 	return flags, nil
-}
-
-// GithubAddCmdFlags - Struct with flags used by githubAddCmd
-type GithubAddCmdFlags struct {
-	GithubOwner string
-	GithubUser  string
-	GithubOrg   string
-	GithubHost  string
-}
-
-func defineGithubCmdFlags(currentCommand *cobra.Command) {
-	currentCommand.Flags().String("github-org", "", "Github Org of repos")
-	currentCommand.Flags().String("github-owner", "", "Github owner of repos")
-	currentCommand.Flags().String("github-host", "github.com", "Github URL")
-	currentCommand.Flags().String("github-user", "", "Github user")
-
-	viper.BindPFlag("github.host", githubAddCmd.Flags().Lookup("github-host"))
-	viper.BindPFlag("github.org", githubAddCmd.Flags().Lookup("github-org"))
-	viper.BindPFlag("github.owner", githubAddCmd.Flags().Lookup("github-owner"))
-	viper.BindPFlag("github.owner", githubAddCmd.Flags().Lookup("github-owner"))
-
-}
-
-func processGithubAddCmdFlags(cmd *cobra.Command) (GithubAddCmdFlags, error) {
-	flags := GithubAddCmdFlags{}
-	user, err := cmd.Flags().GetString("github-user")
-	if err != nil {
-		log.Println("Error Processing - github-user flag")
-		return flags, err
-	}
-	org, err := cmd.Flags().GetString("github-org")
-	if err != nil {
-		log.Println("Error Processing - github-org flag")
-		return flags, err
-	}
-
-	owner, err := cmd.Flags().GetString("github-owner")
-	if err != nil {
-		log.Println("Error Processing - github-owner flag")
-		return flags, err
-	}
-
-	host, err := cmd.Flags().GetString("github-host")
-	if err != nil {
-		log.Println("Error Processing - github-host flag")
-		return flags, err
-	}
-	flags.GithubHost = host
-
-	if owner == "" {
-		if org == "" {
-			owner = user
-		} else {
-			owner = org
-		}
-
-	}
-	flags.GithubOwner = owner
-	flags.GithubOrg = org
-	flags.GithubUser = user
-	return flags, nil
-
 }
